Crate lockedbox

LockedBox

A crate providing an mlock-locked Box<T>, preventing the contents of its memory from being paged to disk.

This crate prevents potential accidental unlocking of memory by ensuring the memory allocated by LockedBox<T> is always a multiple of the operating system’s page size. Because mlock/munlock operate on pages of memory, this guarantees that each LockedBox<T> is guaranteed to have its own lock status.

LockedBox<T> is a thin, safe abstraction built atop memsec.

Alternatives

region

Open-source Licenses

This project, like all projects from Khonsu Labs, is open-source. This repository is available under the MIT License or the Apache License 2.0.

To learn more about contributing, please see CONTRIBUTING.md.

Structs

LockedBox

A [Box]-like type that uses mlock to prevent paging the allocated memory to disk.